app/requirements.txt
Django==3.0.7
Flask-SQLAlchemy==2.4.1
Pillow==5.0.0
asgiref==3.2.10
django-crontab==0.7.1
django-tinymce==3.0.2
psycopg2==2.8.4
pytz==2020.1
sqlparse==0.3.1
django-admin-interface==0.12.3
